gym-planner/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
PyPDF2==3.0.1
reportlab==4.0.9
python-multipart==0.0.6
